Unified-IE working on M1, but it's CPU-only for now, I think
The main thing was making all the dependencies work. Here's the gist:

- Some dependencies had to be unpinned. See the diff.
- grpcio, pyarrow, pandas, numpy and h5py had to be installed with conda
- The mirror used to get the model is broken, so that's gone
- You need to use Pytorch nightly to use MPS, but torchaudio and
  torchvision don't work yet. That doesn't seem to be a problem. Just
  use `conda install pytorch -c pytorch-nightly`, i.e. just pytorch
  without the others. They don't seem to be necessary for finetuning.
- You need to install gnu-grep and gnu-sed from brew and add them to the
  path. That's because some scripts in this repository use flags for sed
  and grep that are not supported by macOS's original versions.

There was also an issue with the code: the parameters for the model and
dataset names weren't working as shown in the README. I've changed the
script to fix that. Also, you need to give the full name of the model
now, i.e. `luyaojie/uie-base-en`.

As far as I know, this should be running in CPU. My next step is to
change this to run using MPS (i.e. the tensor cores in the M1).